merkle_r1cs: the index is a full word everywhere, plus an io_schema
Two halves of the same gap.

`0ff866b` moved the index to a word-aligned position and freed all 128 of
its columns, but the witness generators still wrote only the low `depth`
bits — so the committed word's value was always the bare position. Wiring
a Fiat-Shamir challenge into it would then fail the copy constraint, which
is the one thing the relayout existed to enable. Widen
`ChunkPathInput::index` to `u128` (it IS the word) and write all 128
columns in `build_witness_chunk` too; drop the now-meaningless bound from
`assert_chunk_input`. The packed-vs-bool differential now randomizes the
high half, so the two drivers must agree on every column.

`index_word_high_bits_are_free` pins the mechanism: an opening carrying an
arbitrary challenge word satisfies, folds to the same root, and differs
from the bare-position witness in exactly the index word's high columns —
while the low `depth` stay load-bearing. That is what lets a circuit bind
an opening to a query with no masking gadget: `& (block_len - 1)` is not
computed, it is which columns the relation reads.

Add `MerkleTreeLayout::io_schema` (chunk-leaf only): leaf data, then the
index word, then the two root halves. Siblings are deliberately absent —
they sit unaligned inside each level's padding and are read by no other
gate, and the relation binds them anyway through the root. Circuits pass
them as a `GateType::Hint`.

+++ crates/flock-prover/benches/merkle_l0_opening.rs
@@ -85,7 +85,7 @@ impl Rng {
     fn opening(&mut self, d: usize, leaf: usize) -> ChunkPathInput {
         ChunkPathInput {
             leaf_data: (0..leaf).map(|_| self.next_u32() as u8).collect(),
-            index: ((self.next_u32() as u64) << 32 | self.next_u32() as u64) & ((1u64 << d) - 1),
+            index: ((self.next_u32() as u128) << 32 | self.next_u32() as u128) & ((1u128 << d) - 1),
             siblings: (0..d).map(|_| self.digest()).collect(),
         }
     }
